Separate members into files
